ManuSearch/requirements.txt
python-dotenv
streamlit
streamlit-extras
cachetools
scrapy
trafilatura
json-repair
htmldate
pydantic
requests
openai
asyncio